Fix ODR violation in BaseKeyed serialization
Combined the different implementation of the same tempalte function
specialization into one function.
Register the SerializationHelper only once.
