+++ CondCore/BaseKeyedPlugins/BuildFile.xml
@@ -1,0 +1,5 @@
+<use name="CondCore/ESSources"/>
+<use name="CondFormats/SerializationHelper"/>
+<use name="CondFormats/Calibration"/>
+<use name="CondFormats/DTObjects"/>
+<flags EDM_PLUGIN="1"/>
